item,image경로지정완료(95%완료), sql테이블 itme,item_image변경, list.html수정

+++ Project_test/src/main/java/com/example/demo/image/ItemImage.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class ItemImage {
-    private Long imageNo;    // imageNo는 자동 생성되므로, 객체 생성 시 반드시 포함될 필요 없음
-    private int itemNo;      // 아이템 번호
-    private String imageName; // 이미지 파일 이름
+    private Long imageNo;        // imageNo는 자동 생성되므로, 객체 생성 시 반드시 포함될 필요 없음
+    private Long itemNo;         // 아이템 번호 (Long 타입으로 수정)
+    private String imageName;    // 이미지 파일 이름 (imageName -> filename으로 수정)
 }
